feat: update weight type and add default values for micronutrients - Add migration V5: set default values for micronutrients fields - Add migration V6: change weight type to NUMERIC(5,2) and add status field - Update UserInfo and HistoricoPeso models to use Float for peso - Add status field to HistoricoPeso model - Add default values for micronutrients in MetaNutrientes model

+++ src/main/java/dev/caio/fitsy/model/UserInfo.java
@@ -25,7 +25,7 @@ public class UserInfo {
     private User user;
 
     @Column
-    private Float altura;
+    private Integer altura;
 
     @Column
     private Float peso;
